+++ lib/index.ts
@@ -407,6 +407,27 @@ export default class Qsu {
 		return descending ? result.reverse() : result;
 	}
 
+	static arrGroupByMaxCount(array: any[], maxLengthPerGroup = 1): any[] {
+		const result = [];
+		const arrayLength = array.length;
+		let tempArray = [];
+
+		for (let i = 0; i < arrayLength; i += 1) {
+			if (tempArray.length === maxLengthPerGroup) {
+				result.push(tempArray);
+				tempArray = [];
+			}
+
+			tempArray.push(array[i]);
+		}
+
+		if (tempArray.length > 0) {
+			result.push(tempArray);
+		}
+
+		return result;
+	}
+
 	/*
 	 * Object
 	 * */
@@ -1182,6 +1203,7 @@ export const {
 	arrTo1dArray,
 	sortByObjectKey,
 	sortNumeric,
+	arrGroupByMaxCount,
 	objToQueryString,
 	objToPrettyStr,
 	objFindItemRecursiveByKey,